+++ index.js
@@ -1,11 +1,20 @@
-var getFileType = require('./lib/get-file-type');
-var mapnikAnalyze = require('./lib/mapnik-analyze');
-var tileAnalyze = require('./lib/tile-analyze');
-var reportStats = require('./lib/report-stats');
-var Constants = require('./lib/constants');
+'use strict';
 
+const getFileType = require('./lib/get-file-type');
+const mapnikAnalyze = require('./lib/mapnik-analyze');
+const tileAnalyze = require('./lib/tile-analyze');
+const reportStats = require('./lib/report-stats');
+const Constants = require('./lib/constants');
+
+/**
+ * @param {string} filePath
+ * @param {Object} options
+ * @returns {Object} The stats.
+ */
 function buildGeoStats(filePath, options) {
   options = options || {};
+  options.forceAllAttributes ??= false;
+  options.maxValuesToReport ??= 50;
 
   if (options.attributes) {
     if (options.attributes.length > Constants.ATTRIBUTES_MAX_REPORT) {
@@ -17,7 +26,7 @@ function buildGeoStats(filePath, options) {
   }
 
   return getFileType(filePath)
-    .then(function (fileType) {
+    .then(function(fileType) {
       if (fileType === Constants.FILETYPE_MBTILES) return tileAnalyze(filePath, options);
       return mapnikAnalyze(filePath, fileType, options);
     })