+++ test/test.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const test = require('tap').test;
-const _ = require('lodash');
 const path = require('path');
 const fs = require('fs');
 const sloppySort = require('./utils/sloppy-sort');
@@ -51,9 +50,9 @@ test('Errors when Mapnik-interpreted file not found', t => {
   });
 });
 
-test('GeoJSON with many value types, input matching MBTiles', t => {
+test('GeoJSON with many value types, input matching MBTiles, forceAllAttributes', t => {
   Promise.all([
-    geostats(fixturePath('src/many-types.geojson')),
+    geostats(fixturePath('src/many-types.geojson'), { forceAllAttributes: true }),
     getExpected('many-types-geojson'),
   ]).then((output) => {
     t.same(sloppySort(output[0]), sloppySort(output[1]), 'expected output');
@@ -83,7 +82,7 @@ test('GeoJSON with over 100 unique attributes and values, input matching Shapefi
       t.same(output[0], output[1], 'expected output');
       t.end();
     }).catch(t.threw);
-  }
+  },
 );
 
 // Key difference between the Shapefile and the GeoJSON output right now
@@ -98,7 +97,7 @@ test('Shapefile with over 100 unique attributes and values, input matching GeoJS
       t.same(output[0], output[1], 'expected output');
       t.end();
     }).catch(t.threw);
-  }
+  },
 );
 
 // Key difference between the CSV and Shapefile and GeoJSON is that it
@@ -112,18 +111,8 @@ test('CSV with over 100 unique attributes and values, input matching GeoJSON and
       t.same(output[0], output[1], 'expected output');
       t.end();
     }).catch(t.threw);
-  }
+  },
 );
-
-test('Shapefile with over 1000 unique values', t => {
-  Promise.all([
-    geostats(fixturePath('src/ports/ports.shp')),
-    getExpected('ports'),
-  ]).then((output) => {
-    t.same(output[0], output[1], 'expected output');
-    t.end();
-  }).catch(t.threw);
-});
 
 test('MBTiles with gzipped data', t => {
   Promise.all([
@@ -273,20 +262,15 @@ test('invalid file format', t => {
   });
 });
 
-test('Shapefile with specified attribute with over 1000 values', t => {
+test('GeoJSON with specified name attribute', t => {
   Promise.all([
-    geostats(fixturePath('src/ports/ports.shp'), {
+    geostats(fixturePath('src/many-types.geojson'), {
       attributes: ['name'],
     }),
-    getExpected('ports-only-name'),
+    getExpected('many-types-geojson-name-only'),
   ]).then((output) => {
     const actual = sloppySort(output[0]);
     t.same(actual, sloppySort(output[1]), 'expected output');
-    const nameAttribute = _.find(actual.layers[0].attributes, attribute => {
-      return attribute.attribute === 'name';
-    });
-    t.equal(nameAttribute.count, 1042, 'value count did not stop at 1000');
-    t.equal(nameAttribute.values.length, 1042, 'value details did not stop at 100');
     t.end();
   }).catch(t.threw);
 });
@@ -303,81 +287,26 @@ test('GeoJSON with specified attributes', t => {
   }).catch(t.threw);
 });
 
-test('GeoJSON with over 10000 unique values and no specified attributes', t => {
-  Promise.all([
-    geostats(fixturePath('src/myriad-values.geojson')),
-    getExpected('myriad-values-all-attrs'),
-  ]).then((output) => {
-    const actual = sloppySort(output[1]);
-    t.same(sloppySort(output[0]), actual, 'expected output');
-    t.ok(actual.layers[0].attributes.every(attribute => {
-      return attribute.count === 1000;
-    }), 'value counts stop at 1000');
-    t.ok(actual.layers[0].attributes.every(attribute => {
-      return attribute.values.length === 100;
-    }), 'value details stop at 100');
-    t.end();
-  }).catch(t.threw);
-});
-
-test('GeoJSON with over 10000 unique values and one specified attribute', t => {
-  Promise.all([
-    geostats(fixturePath('src/myriad-values.geojson'), {
-      attributes: ['prop-3'],
-    }),
-    getExpected('myriad-values-1-attr'),
-  ]).then((output) => {
-    const actual = sloppySort(output[1]);
-    t.same(sloppySort(output[0]), actual, 'expected output');
-    t.ok(actual.layers[0].attributes.every(attribute => {
-      return attribute.count === 10010;
-    }), 'value count does not stop yet');
-    t.ok(actual.layers[0].attributes.every(attribute => {
-      return attribute.values.length === 10000;
-    }), 'value details stop at 10000');
-    t.end();
-  }).catch(t.threw);
-});
-
-test('GeoJSON with over 10000 unique values and five specified attribute', t => {
-  Promise.all([
-    geostats(fixturePath('src/myriad-values.geojson'), {
-      attributes: ['prop-1', 'prop-2', 'prop-3', 'prop-4', 'prop-5'],
-    }),
-    getExpected('myriad-values-5-attrs'),
-  ]).then((output) => {
-    const actual = sloppySort(output[1]);
-    t.same(sloppySort(output[0]), actual, 'expected output');
-    t.ok(actual.layers[0].attributes.every(attribute => {
-      return attribute.count === 10010;
-    }), 'value counts does not stop yet');
-    t.ok(actual.layers[0].attributes.every(attribute => {
-      return attribute.values.length === 2000;
-    }), 'value details stop at 2000');
-    t.end();
-  }).catch(t.threw);
-});
-
 test('Trying to report on more than 100 attributes', t => {
   t.throws(() => {
     geostats(fixturePath('src/populations-plus.geojson'), {
       attributes: ['attr-0', 'attr-1', 'attr-2', 'attr-3', 'attr-4', 'attr-5',
-      'attr-6', 'attr-7', 'attr-8', 'attr-9', 'attr-10', 'attr-11', 'attr-12',
-      'attr-13', 'attr-14', 'attr-15', 'attr-16', 'attr-17', 'attr-18', 'attr-19',
-      'attr-20', 'attr-21', 'attr-22', 'attr-23', 'attr-24', 'attr-25', 'attr-26',
-      'attr-27', 'attr-28', 'attr-29', 'attr-30', 'attr-31', 'attr-32', 'attr-33',
-      'attr-34', 'attr-35', 'attr-36', 'attr-37', 'attr-38', 'attr-39', 'attr-40',
-      'attr-41', 'attr-42', 'attr-43', 'attr-44', 'attr-45', 'attr-46', 'attr-47',
-      'attr-48', 'attr-49', 'attr-50', 'attr-51', 'attr-52', 'attr-53', 'attr-54',
-      'attr-55', 'attr-56', 'attr-57', 'attr-58', 'attr-59', 'attr-60', 'attr-61',
-      'attr-62', 'attr-63', 'attr-64', 'attr-65', 'attr-66', 'attr-67', 'attr-68',
-      'attr-69', 'attr-70', 'attr-71', 'attr-72', 'attr-73', 'attr-74', 'attr-75',
-      'attr-76', 'attr-77', 'attr-78', 'attr-79', 'attr-80', 'attr-81', 'attr-82',
-      'attr-83', 'attr-84', 'attr-85', 'attr-86', 'attr-87', 'attr-88', 'attr-89',
-      'attr-90', 'attr-91', 'attr-92', 'attr-93', 'attr-94', 'attr-95', 'attr-96',
-      'attr-97', 'attr-98', 'attr-99', 'attr-100', 'attr-101', 'attr-102',
-      'attr-103', 'attr-104', 'attr-105', 'attr-106', 'attr-107', 'attr-108',
-      'attr-109'],
+        'attr-6', 'attr-7', 'attr-8', 'attr-9', 'attr-10', 'attr-11', 'attr-12',
+        'attr-13', 'attr-14', 'attr-15', 'attr-16', 'attr-17', 'attr-18', 'attr-19',
+        'attr-20', 'attr-21', 'attr-22', 'attr-23', 'attr-24', 'attr-25', 'attr-26',
+        'attr-27', 'attr-28', 'attr-29', 'attr-30', 'attr-31', 'attr-32', 'attr-33',
+        'attr-34', 'attr-35', 'attr-36', 'attr-37', 'attr-38', 'attr-39', 'attr-40',
+        'attr-41', 'attr-42', 'attr-43', 'attr-44', 'attr-45', 'attr-46', 'attr-47',
+        'attr-48', 'attr-49', 'attr-50', 'attr-51', 'attr-52', 'attr-53', 'attr-54',
+        'attr-55', 'attr-56', 'attr-57', 'attr-58', 'attr-59', 'attr-60', 'attr-61',
+        'attr-62', 'attr-63', 'attr-64', 'attr-65', 'attr-66', 'attr-67', 'attr-68',
+        'attr-69', 'attr-70', 'attr-71', 'attr-72', 'attr-73', 'attr-74', 'attr-75',
+        'attr-76', 'attr-77', 'attr-78', 'attr-79', 'attr-80', 'attr-81', 'attr-82',
+        'attr-83', 'attr-84', 'attr-85', 'attr-86', 'attr-87', 'attr-88', 'attr-89',
+        'attr-90', 'attr-91', 'attr-92', 'attr-93', 'attr-94', 'attr-95', 'attr-96',
+        'attr-97', 'attr-98', 'attr-99', 'attr-100', 'attr-101', 'attr-102',
+        'attr-103', 'attr-104', 'attr-105', 'attr-106', 'attr-107', 'attr-108',
+        'attr-109'],
     });
   }, 'throws');
   t.end();